+++ script/Deploy.s.sol
@@ -5,10 +5,10 @@ import { Script } from "forge-std/Script.sol";
 
 import { WstETH } from "test/utils/WstEth.sol";
 
-import { IOracleMiddleware } from "src/interfaces/IOracleMiddleware.sol";
+import { IOracleMiddleware } from "src/interfaces/OracleMiddleware/IOracleMiddleware.sol";
 import { Usdn } from "src/Usdn.sol";
 import { UsdnProtocol } from "src/UsdnProtocol/UsdnProtocol.sol";
-import { OracleMiddleware } from "../src/OracleMiddleware/OracleMiddleware.sol";
+import { WstEthOracleMiddleware } from "src/OracleMiddleware/WstEthOracleMiddleware.sol";
 
 contract Deploy is Script {
     function run() external {
@@ -34,14 +34,14 @@ contract Deploy is Script {
 
         // Deploy Oracle middleware if needed
         address middlewareAddress = vm.envOr("MIDDLEWARE_ADDRESS", address(0));
-        IOracleMiddleware middleware;
+        WstEthOracleMiddleware middleware;
         if (middlewareAddress != address(0)) {
-            middleware = IOracleMiddleware(middlewareAddress);
+            middleware = WstEthOracleMiddleware(middlewareAddress);
         } else {
             address pythAddress = vm.envAddress("PYTH_ADDRESS");
-            bytes32 pythPriceId = vm.envBytes32("PYTH_WSTETH_PRICE_ID");
+            bytes32 pythPriceId = vm.envBytes32("PYTH_STETH_PRICE_ID");
             address chainlinkPriceAddress = vm.envAddress("CHAINLINK_STETH_PRICE_ADDRESS");
-            middleware = new OracleMiddleware(pythAddress, pythPriceId, chainlinkPriceAddress);
+            middleware = new WstEthOracleMiddleware(pythAddress, pythPriceId, chainlinkPriceAddress, wstETHAddress);
             middlewareAddress = address(middleware);
         }