+++ test/unit/OracleMiddleware/utils/Constants.sol
@@ -7,3 +7,5 @@ pragma solidity 0.8.20;
 
 uint256 constant ETH_PRICE = 2000 * 1e8;
 uint256 constant ETH_CONF = 20 * 1e8;
+uint256 constant STETH_PRICE = 2000 * 1e8;
+uint256 constant STETH_CONF = 20 * 1e8;