+++ test/unit/USDN/Usdn.fuzzing.t.sol
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.20;
 
-import { Vm } from "forge-std/Vm.sol";
-
 import { UsdnTokenFixture } from "test/unit/USDN/utils/Fixtures.sol";
 import { USER_1 } from "test/utils/Constants.sol";